Latest build

+++ command.c
@@ -40,7 +40,7 @@ int g_debug;
  *
  */
 
-void append_background_process(command_t *command, size_t pid) {
+void append_background_process(command_t *command, int pid) {
   if (g_background_process_head == NULL) {
     g_background_process_head = (node_t *) malloc(sizeof(node));
     g_background_process_head->next = NULL;
